Create a separate class for compactgraph
